Part 11 of Final
Adding a checkpoint at the beginning of the cave. Added water that will have a hidden hazard at the bottom of it. Added a poison trap on the other end of the maze. Added a treasure chest in the other part of chest. Added the post process volume needs effects added. Added a Global PPV needs added effects.